Refactor stub implementation of LTI Provider. BLD-601.
