+++ common/djangoapps/terrain/start_stubs.py
@@ -6,11 +6,12 @@ from lettuce import before, after, world
 from django.conf import settings
 from terrain.stubs.youtube import StubYouTubeService
 from terrain.stubs.xqueue import StubXQueueService
-
+from terrain.stubs.lti import StubLtiService
 
 SERVICES = {
     "youtube": {"port": settings.YOUTUBE_PORT, "class": StubYouTubeService},
     "xqueue": {"port": settings.XQUEUE_PORT, "class": StubXQueueService},
+    "lti": {"port": settings.LTI_PORT, "class": StubLtiService},
 }
