+++ lms/envs/devstack.py
@@ -7,6 +7,7 @@ from .aws import * # pylint: disable=wildcard-import, unused-wildcard-import
 DEBUG = True
 USE_I18N = True
 TEMPLATE_DEBUG = True
+SITE_NAME = 'localhost:8000'
 # By default don't use a worker, execute tasks as if they were local functions
 CELERY_ALWAYS_EAGER = True
